// types/applicant.ts
export interface Candidato {
    id: string
    nome: string
    cognome: string
    email: string
    telefono: string
    data_nascita: string
    residenza: string
    domiciliazione: string
    universita: string
    facolta: string
    corso: string
    tipo_curriculum: string
    anno_corso: string
    area_1: string
    area_2: string
    come_conosciuto_jesap: string
    perche_jesap: string
    perche_area: string
    conosce_membri_jesap: string
    in_je_italy: string
    resume_url: string // URL per il download del curriculum
    created_at: string
  }
  
  